Finish reservation screen

+++ AdminMainWindow.py
@@ -1,11 +1,14 @@
 import sys
-from PyQt5.QtWidgets import QApplication, QMainWindow, QTabWidget, QWidget, QVBoxLayout, QTableWidget, QTableWidgetItem, QAbstractItemView, QHeaderView, QPushButton, QHBoxLayout, QMessageBox
+from PyQt5.QtWidgets import QApplication, QMainWindow, QTabWidget, QWidget, QVBoxLayout, QTableWidget, QTableWidgetItem, QAbstractItemView, QHeaderView, QPushButton, QHBoxLayout, QMessageBox, QLabel, QDateEdit
 from PyQt5.QtGui import QFont, QIcon
+from PyQt5.QtCore import QDate
 from AddClientScreen import AddClientScreen
 from EditClientScreen import EditClientScreen
 from AddRoomScreen import AddRooomScreen
 from EditRoomScreen import EditRoomScreen
 from AddReservationScreen import AddReservationScreen
+from datetime import datetime
+from EditReservationScreen import EditReservationScreen
 
 class AdminMainWindow(QMainWindow):
     def __init__(self, conn, isAdmin):
@@ -48,7 +51,30 @@ class AdminMainWindow(QMainWindow):
 
         #Room
         tab2 = QWidget()
+
         tab2_layout = QVBoxLayout()
+
+        date_layout = QHBoxLayout()
+        
+        checkin_label = QLabel("Дата въезда:")
+        self.checkin_date = QDateEdit()
+        self.checkin_date.setCalendarPopup(True)
+        self.checkin_date.setDateRange(QDate.currentDate(), QDate(2100, 12, 31))
+        date_layout.addWidget(checkin_label)
+        date_layout.addWidget(self.checkin_date)
+        
+        checkout_label = QLabel("Дата выезда:")
+        self.checkout_date = QDateEdit()
+        self.checkout_date.setCalendarPopup(True)
+        self.checkout_date.setDateRange(QDate.currentDate(), QDate(2100, 12, 31))
+        date_layout.addWidget(checkout_label)
+        date_layout.addWidget(self.checkout_date)
+
+        self.search_button = QPushButton('Найти', self)
+        date_layout.addWidget(self.search_button)
+           
+        tab2_layout.addLayout(date_layout)
+
         self.room_table = QTableWidget()
         self.room_table.setSelectionBehavior(QAbstractItemView.SelectRows)
         self.room_table.setSelectionMode(QAbstractItemView.SingleSelection)  
@@ -139,6 +165,8 @@ class AdminMainWindow(QMainWindow):
 
         self.setCentralWidget(tab_widget)
 
+        self.search_button.clicked.connect(self.onSearchButtonClicked)
+
         add_client_btn.clicked.connect(lambda: self.add_row(self.client_table))
         edit_client_btn.clicked.connect(lambda: self.edit_row(self.client_table))
         delete_client_btn.clicked.connect(lambda: self.delete_row(self.client_table))
@@ -177,6 +205,36 @@ class AdminMainWindow(QMainWindow):
 
     def enableAdminMainWindow(self):
         self.setEnabled(True)
+
+    def onSearchButtonClicked(self):
+        checkin_date = self.checkin_date.date().toString("yyyy-MM-dd")
+        checkout_date = self.checkout_date.date().toString("yyyy-MM-dd")
+        current_date = datetime.now().strftime("%Y-%m-%d")
+
+        print(type(checkin_date))
+
+        if (checkin_date >= checkout_date) or (checkin_date < current_date):  
+            QMessageBox.warning(self, "Ошибка", "Некорректно выбраны даты")
+        else:
+            with self.conn as conn:
+                with conn.cursor() as cursor:
+                    cursor.execute("""
+                    SELECT * FROM get_available_rooms_query(%s, %s);
+                    """, (checkin_date, checkout_date))
+                    rows = cursor.fetchall()
+                    
+                    self.room_table.setRowCount(len(rows))
+                    self.room_table.setColumnCount(len(rows[0]))
+
+                    column_labels = ['ID комнаты','Комната', 'Вместимость', 'Цена']  
+                    self.room_table.setHorizontalHeaderLabels(column_labels)
+                    
+                    for i, row in enumerate(rows):
+                        for j, value in enumerate(row):
+                            item = QTableWidgetItem(str(value))
+                            item.setFont(QFont("Arial", 10)) 
+                            self.room_table.setItem(i, j, item)
+                    self.room_table.setColumnHidden(0, True)       
 
     def add_row(self, table_widget):
         if (table_widget == self.client_table):
@@ -217,6 +275,20 @@ class AdminMainWindow(QMainWindow):
                 self.edit_room_screen = EditRoomScreen(self.conn, room_id, room, capacity, payment, self)
                 self.edit_room_screen.show()
                 self.setDisabled(True)
+
+        elif (table_widget == self.reservation_table):
+            selected_row = table_widget.currentRow()
+            if selected_row != -1:
+                reservation_id = table_widget.item(selected_row, 0).text()
+                client_id = table_widget.item(selected_row, 1).text()
+                client_data = table_widget.item(selected_row, 2).text()
+                room = table_widget.item(selected_row, 3).text()
+                check_id_date = table_widget.item(selected_row, 5).text()
+                check_out_date = table_widget.item(selected_row, 6).text()
+
+                self.edit_reservation_screen = EditReservationScreen(self.conn, reservation_id, client_data, client_id, room, check_id_date, check_out_date, self)
+                self.edit_reservation_screen.show()
+                self.setDisabled(True)
         
     def delete_row(self, table_widget):
         if (table_widget == self.client_table):
@@ -226,7 +298,7 @@ class AdminMainWindow(QMainWindow):
                 try:
                     with self.conn as conn:
                         with conn.cursor() as cursor:
-                            cursor.execute("SELECT delete_client(%s)", (client_id))
+                            cursor.execute("SELECT delete_client(%s)", (client_id,))
                             conn.commit()
                             QMessageBox.information(self, 'Успех', f'Клиент удален!')
                 except Exception as e:
@@ -239,38 +311,57 @@ class AdminMainWindow(QMainWindow):
                 try:
                     with self.conn as conn:
                         with conn.cursor() as cursor:
-                            cursor.execute("SELECT delete_room(%s)", (room_id))
+                            cursor.execute("SELECT delete_room(%s)", (room_id,))
                             conn.commit()
                             QMessageBox.information(self, 'Успех', f'Номер удален!')
                 except Exception as e:
                     QMessageBox.warning(self, 'Ошибка', f'Не удалось удалить номер: ')
                 self.refresh_tables() 
+        elif (table_widget == self.reservation_table):
+            selected_row = table_widget.currentRow()
+            if selected_row != -1:
+                reservation_id = table_widget.item(selected_row, 0).text()
+                try:
+                    with self.conn as conn:
+                        with conn.cursor() as cursor:
+                            cursor.execute("SELECT delete_reservation(%s)", (reservation_id,))
+                            conn.commit()
+                            QMessageBox.information(self, 'Успех', f'Бронирование удалено')
+                except Exception as e:
+                    QMessageBox.warning(self, 'Ошибка', f'Не удалось удалить бронирование: {str(e)}')
+                self.refresh_tables() 
+
 
     def display_data(self, table_widget, query):
         with self.conn as conn:
             with conn.cursor() as cursor:
                 cursor.execute(query)
                 rows = cursor.fetchall()
-                
                 table_widget.setRowCount(len(rows))
-                table_widget.setColumnCount(len(rows[0]))
+
+                column_count = cursor.description
+                num_columns = len(column_count)
+                table_widget.setColumnCount(num_columns)
+ 
                 
                 if (table_widget == self.client_table):
                     column_labels = ['ID клиента','Фамилия', 'Имя', 'Номер телефона', 'Почта']  
                 elif (table_widget == self.room_table):
                     column_labels = ['ID комнаты','Комната', 'Вместимость', 'Цена']  
                 elif (table_widget == self.reservation_table):
-                    column_labels = ['ID брони','Клиент', 'Комната', 'Дата бронированя', 'Дата заселения', 'Дата выселения']  
+                    column_labels = ['ID брони', 'ID клиента', 'Клиент', 'Комната', 'Дата бронированя', 'Дата заселения', 'Дата выселения']  
                 else:
                     column_labels = ['ID оплаты','Тип оплаты', 'Номер карты', 'Код карты', 'Месяц конца обс.', 'Год конца обс.', 'Номер брони', 'Сумма']  
                 table_widget.setHorizontalHeaderLabels(column_labels)
-                
-                for i, row in enumerate(rows):
-                    for j, value in enumerate(row):
-                        item = QTableWidgetItem(str(value))
-                        item.setFont(QFont("Arial", 10)) 
-                        table_widget.setItem(i, j, item)
-                table_widget.setColumnHidden(0, True)     
+
+                if (len(rows) != 0):          
+                    for i, row in enumerate(rows):
+                        for j, value in enumerate(row):
+                            item = QTableWidgetItem(str(value))
+                            item.setFont(QFont("Arial", 10)) 
+                            table_widget.setItem(i, j, item)
+                    table_widget.setColumnHidden(0, True)     
+                    table_widget.setColumnHidden(1, True)     
 
 if __name__ == '__main__':
     app = QApplication(sys.argv)
